[FIX] Update next.config.js, add million.js for speed, organize files, do a little magic
