--- site/components/Map/index.js
@@ -1,5 +1,0 @@
-import dynamic from "next/dynamic";
-
-const Map = dynamic(() => import("./Map"), { ssr: false });
-
-export default Map;
