+++ site/components/heading.js
@@ -57,8 +57,31 @@ export const DirectoryVideoSection = () => (
         Clubs Directory
       </Heading>
       <Text variant="lead">
-        Hack Together, Club Together because We're Better Together
+        Hack Together, Club Together because <strong>We're Better Together</strong>
       </Text>
     </Container>
   </Box>
 );
+
+export const DirectoryHeading = () => (
+  <Container>
+    <Heading
+      as="h2"
+      variant="title"
+      sx={{
+        color: "red",
+        pt: [3, 4],
+        pb: [3, 4],
+        span: {
+          WebkitTextStroke: "currentColor",
+          WebkitTextStrokeWidth: ["1.5px", "2.5px"],
+          WebkitTextFillColor: "transparent",
+        },
+      }}
+    >
+      Clubs Directory unlocks the power of <span>cross-club collaboration</span>
+      , empowering clubs to transcend boundaries and hack together.
+    </Heading>
+  </Container>
+);
+
