+++ site/components/map-components/index.js
@@ -1,0 +1,5 @@
+import dynamic from "next/dynamic";
+
+const Map = dynamic(() => import("@/components/map"), { ssr: false });
+
+export default Map;
