+++ site/components/nav.js
@@ -1,39 +1,21 @@
-import {
-  Box,
-  Container,
-  IconButton,
-  Image,
-  Link as A,
-  useColorMode,
-} from "theme-ui";
-import { useRouter } from "next/router";
+import { Box, Container, Image } from "theme-ui";
 import Link from "next/link";
 
-
 const Flag = () => (
-  <Box
-  onClick={() => 
-  {
-    window.open("https://hackclub.com/", "_blank");
-  }}
-    href="https://hackclub.com/"
-    target="_blank"
-    rel="noopener noreferrer"
-    aria-label="Hack Club homepage"
-    sx={{ mt: -3,}}
-  >
-    <Image
-      src="https://assets.hackclub.com/flag-orpheus-top.svg"
-      alt="Hack Club flag"
-      sx={{ width: [96, 128] }}
-    />
-  </Box>
+  <Link href="https://hackclub.com/" target="_blank">
+    <Box aria-label="Hack Club Homepage" sx={{ mt: -3 }}>
+      <Image
+        src="https://assets.hackclub.com/flag-orpheus-top.svg"
+        alt="Hack Club flag"
+        sx={{ width: [96, 128] }}
+      />
+    </Box>
+  </Link>
 );
-
 
 export default () => {
   return (
-    <Container style={{position: "absolute", zIndex: 1, cursor: "pointer"}}>
+    <Container style={{ position: "absolute", zIndex: 1, cursor: "pointer" }}>
       <Box
         as="nav"
         sx={{
