+++ site/lib/data.js
@@ -1,0 +1,9 @@
+export const continents = [
+    "Asia",
+    "North America",
+    "Europe",
+    "South America",
+    "Africa",
+    "Australia/Oceania",
+    "Antarctica",
+  ];
