+++ site/next.config.js
@@ -1,2 +1,5 @@
-const withMDX = require("@next/mdx")({ extension: /\.mdx?$/ });
-module.exports = withMDX({ pageExtensions: ["js", "jsx", "mdx"] });
+const million = require("million/compiler");
+/** @type {import('next').NextConfig} */
+const nextConfig = {};
+
+module.exports = million.next(nextConfig, { auto: { rsc: true } });
